Move homepage logic to component and use routing config for homepage to render
For the hotel/city/country detail pages, I will need to hose more pages.  This commit is the
foundation for more pages

I added a crude navbar to route back to home, ready for the new pages

Using the API url util
